+++ src/DynamicQueryable/DynamicQueryable.Aggregate.cs
@@ -70,38 +70,6 @@ namespace System.Linq.Dynamic {
             return ExecuteOptionalExpression(source, "Min", selector, true, variables, values);
         }
 
-        public static int Count<T>(this IQueryable<T> source, string predicate, params object[] values) {
-            return Count(source, predicate, null, values);
-        }
-
-        public static int Count<T>(this IQueryable<T> source, string predicate, IDictionary<string, object> variables, params object[] values) {
-            return Count((IQueryable)source, predicate, variables, values);
-        }
-
-        public static int Count(this IQueryable source, string predicate = null, params object[] values) {
-            return Count(source, predicate, null, values);
-        }
-
-        public static int Count(this IQueryable source, string predicate, IDictionary<string, object> variables, params object[] values) {
-            return (int)ExecuteOptionalExpression(source, "Count", predicate, string.IsNullOrEmpty(predicate), variables, values);
-        }
-
-        public static long LongCount<T>(this IQueryable<T> source, string predicate, params object[] values) {
-            return LongCount(source, predicate, null, values);
-        }
-
-        public static long LongCount<T>(this IQueryable<T> source, string predicate, IDictionary<string, object> variables, params object[] values) {
-            return LongCount((IQueryable)source, predicate, variables, values);
-        }
-
-        public static long LongCount(this IQueryable source, string predicate = null, params object[] values) {
-            return LongCount(source, predicate, null, values);
-        }
-
-        public static long LongCount(this IQueryable source, string predicate, IDictionary<string, object> variables, params object[] values) {
-            return (long)ExecuteOptionalExpression(source, "LongCount", predicate, string.IsNullOrEmpty(predicate), variables, values);
-        }
-
         public static object Aggregate(this IQueryable source, string func, params object[] values) {
             return Aggregate(source, func, null, values);
         }
